Use a virtual environment for minio_bucket.py
Instead of messing around with the global system python installation we
can just use a virtual environment to execute the miniio_bucket.py
script.

- Added a new variable (minio_venv_path),
- changed the shebang in minio_bucket.py to make it accept our virtual
  environment,
- install a dependency for virtualenv,
- install pip dependencies to virtual environment,
- prepend virtual environment bin path to minio_bucket.py PATH to make
  it use the virtual environment

This is a bit hacky but since we will only be using this for
minio_bucket.py invoked by Ansible I thought it was acceptable.  The
virtual environment uses approx. 22 MB of storage on my server for this.